requirements.txt
flask==3.1.0
flask-swagger-ui==4.11.0
psycopg2-binary==2.9.10